Use classes for entities

+++ src/app/actions.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { revalidatePath } from "next/cache";
-import { BeerOrder, Entity } from "../db/models";
+import { BeerOrder } from "../db/models";
 import { store } from "../db/store";
 
 export async function addOrderAction(formData: FormData) {
@@ -10,31 +10,30 @@ export async function addOrderAction(formData: FormData) {
 
   const session = store.openSession();
 
-  await session.store<BeerOrder>({
-    id: "",
-    collection: "beerOrders",
-    beerType,
-    liters,
-    isDone: false,
-    createDate: new Date(),
-  });
+  await session.store<BeerOrder>(new BeerOrder(null, beerType, liters));
   await session.saveChanges();
 
-  session.dispose();
   revalidatePath("/");
 }
 
-export async function deleteOrderAction(id: Entity["id"]) {
+export async function deleteOrderAction(id: BeerOrder["id"]) {
+  if (!id) {
+    throw new Error("ID is required")
+  };
+
   const session = store.openSession();
 
   await session.delete<BeerOrder>(id);
   await session.saveChanges();
 
-  session.dispose();
   revalidatePath("/");
 }
 
-export async function toggleOrderAction(id: Entity["id"]) {
+export async function toggleOrderAction(id: BeerOrder["id"]) {
+  if (!id) {
+    throw new Error("ID is required")
+  };
+
   const session = store.openSession();
 
   const order = await session.load<BeerOrder>(id);
@@ -45,6 +44,5 @@ export async function toggleOrderAction(id: Entity["id"]) {
   order.isDone = !order.isDone;
   await session.saveChanges();
 
-  session.dispose();
   revalidatePath("/");
 }
